api/requirements.txt
amqp==2.1.3
aniso8601==1.2.0
anyjson==0.3.3
asgi-redis==1.0.0
asgiref==1.0.0
autobahn==0.17.0
billiard==3.5.0.2
blinker==1.4
block-disposable-email==1.0.1
boto==2.44.0
botocore==1.4.63
celery==4.0.1
channels==0.17.3
click==6.6
colorama==0.3.7
constantly==15.1.0
cycler==0.10.0
daphne==0.15.0
dask==0.12.0
decorator==4.0.10
Django==1.10.4
django-autoslug==1.9.3
django-cors-headers==1.1.0
django-disposable-email-checker==1.2.1
django-extensions==1.7.1
django-filter==0.13.0
django-graphiql==0.4.4
django-jsonfield==1.0.1
git+https://github.com/vwrobel/django-jwt-auth.git
django-storages==1.5.1
django-webpack-loader==0.3.3
djangorestframework==3.4.3
djangorestframework-csv==2.0.0
djangorestframework-jwt==1.8.0
docutils==0.12
ez-setup==0.9
filterpy==0.1.5
Flask==0.11.1
Flask-RESTful==0.3.5
futures==3.0.5
gitdb==0.6.4
GitPython==2.0.8
graphene==1.0
graphene-django==1.0
graphql-core==1.0
graphql-django-view==1.3
graphql-relay==0.4.4
image==1.5.3
imutils==0.3.6
incremental==16.10.1
iso8601==0.1.11
itsdangerous==0.24
Jinja2==2.8
jmespath==0.9.0
kombu==4.0.1
Markdown==2.6.6
MarkupSafe==0.23
matplotlib==1.5.3
msgpack-python==0.4.8
networkx==1.11
nose==1.3.7
numpy==1.11.2
pafy==0.5.1
pandas==0.19.1
Pillow==3.4.2
promise==1.0.0
psycopg2==2.6.2
py==1.4.31
pyasn1==0.1.9
PyJWT==1.4.2
pyparsing==2.1.10
pytest==3.0.4
python-dateutil==2.5.3
python-dotenv==0.6.1
pytz==2016.10
PyYAML==3.12
redis==2.10.5
requests==2.10.0
rest-framework-auth0==0.1.1
rest-pandas==0.5.0
rsa==3.4.2
scikit-image==0.12.3
scipy==0.18.1
singledispatch==3.4.0.3
six==1.10.0
smmap==0.9.0
toolz==0.8.0
tqdm==4.8.3
Twisted==16.6.0
txaio==2.5.2
typing==3.5.2.2
unicodecsv==0.14.1
vine==1.1.3
webcolors==1.5
Werkzeug==0.11.11
zope.interface==4.3.2
